moving bin functions from plamsadatahandling to bin
